+++ src/flows/context/query.tsx
@@ -1,5 +1,6 @@
 import React, {FC, useContext, useMemo, useEffect} from 'react'
 import {useDispatch, useSelector} from 'react-redux'
+import {parse} from 'src/external/parser'
 import {runQuery} from 'src/shared/apis/query'
 import {getWindowVars} from 'src/variables/utils/getWindowVars'
 import {buildVarsOption} from 'src/variables/utils/buildVarsOption'
@@ -19,7 +20,6 @@ import {
   generateHashedQueryID,
   setQueryByHashID,
 } from 'src/timeMachine/actions/queries'
-import {findOrgID} from 'src/flows/shared/utils'
 import {notify} from 'src/shared/actions/notifications'
 
 // Constants
@@ -53,6 +53,44 @@ export const QueryContext = React.createContext<QueryContextType>(
 )
 
 const PREVIOUS_REGEXP = /__PREVIOUS_RESULT__/g
+
+const findOrgID = (text, buckets) => {
+  const ast = parse(text)
+
+  const _search = (node, acc = []) => {
+    if (!node) {
+      return acc
+    }
+    if (
+      node?.type === 'CallExpression' &&
+      node?.callee?.type === 'Identifier' &&
+      node?.callee?.name === 'from' &&
+      node?.arguments[0]?.properties[0]?.key?.name === 'bucket'
+    ) {
+      acc.push(node)
+    }
+
+    Object.values(node).forEach(val => {
+      if (Array.isArray(val)) {
+        val.forEach(_val => {
+          _search(_val, acc)
+        })
+      } else if (typeof val === 'object') {
+        _search(val, acc)
+      }
+    })
+
+    return acc
+  }
+
+  const queryBuckets = _search(ast).map(
+    node => node?.arguments[0]?.properties[0]?.value.value
+  )
+
+  const bucket = buckets.find(buck => queryBuckets.includes(buck.name))
+
+  return bucket?.orgID
+}
 
 export const QueryProvider: FC = ({children}) => {
   const {flow} = useContext(FlowContext)
@@ -141,7 +179,7 @@ export const QueryProvider: FC = ({children}) => {
       })
   }
 
-  const query = (text: string) => {
+  const query = (text: string): Promise<FluxResult> => {
     const orgID = findOrgID(text, buckets)
     const windowVars = getWindowVars(text, vars)
     const extern = buildVarsOption([...vars, ...windowVars])
@@ -164,7 +202,7 @@ export const QueryProvider: FC = ({children}) => {
           raw: raw.csv,
           parsed: fromFlux(raw.csv),
           error: null,
-        }
+        } as FluxResult
       })
   }
